move spec, add test for merge hf with block num
