+++ packages/common/tests/hardforks.spec.ts
@@ -63,75 +63,6 @@ tape('[Common]: Hardfork logic', function (t: tape.Test) {
     st.end()
   })
 
-  t.test('test post merge hardforks using Sepolia', function (st: tape.Test) {
-    const c = new Common({ chain: Chain.Sepolia })
-    let msg = 'should get HF correctly'
-
-    st.equal(c.getHardforkByBlockNumber(0), Hardfork.London, msg)
-    //using Hardfork.London even though Merge is expected after terminal block 1450408
-    st.equal(c.getHardforkByBlockNumber(1450409), Hardfork.London, msg)
-    st.equal(c.getHardforkByBlockNumber(1450409, BigInt('17000000000000000')), Hardfork.Merge, msg)
-    st.equal(c.getHardforkByBlockNumber(1735371), Hardfork.MergeForkIdTransition, msg)
-    st.equal(
-      c.getHardforkByBlockNumber(1735371, BigInt('17000000000000000')),
-      Hardfork.MergeForkIdTransition,
-      msg
-    )
-    try {
-      st.equal(
-        c.getHardforkByBlockNumber(1735371, BigInt('15000000000000000')),
-        Hardfork.MergeForkIdTransition,
-        msg
-      )
-      st.fail('should have thrown as specified td < merge ttd for a post merge hardfork')
-    } catch (error) {
-      st.pass('throws error as specified td < merge ttd for a post merge hardfork')
-    }
-
-    msg = 'should set HF correctly'
-
-    st.equal(c.setHardforkByBlockNumber(0), Hardfork.London, msg)
-    //using Hardfork.London even though Merge is expected after terminal block 1450408
-    st.equal(c.setHardforkByBlockNumber(1450409), Hardfork.London, msg)
-    st.equal(c.setHardforkByBlockNumber(1450409, BigInt('17000000000000000')), Hardfork.Merge, msg)
-    st.equal(c.setHardforkByBlockNumber(1735371), Hardfork.MergeForkIdTransition, msg)
-    st.equal(
-      c.setHardforkByBlockNumber(1735371, BigInt('17000000000000000')),
-      Hardfork.MergeForkIdTransition,
-      msg
-    )
-    try {
-      st.equal(
-        c.setHardforkByBlockNumber(1735371, BigInt('15000000000000000')),
-        Hardfork.MergeForkIdTransition,
-        msg
-      )
-      st.fail('should have thrown as specified td < merge ttd for a post merge hardfork')
-    } catch (error) {
-      st.pass('throws error as specified td < merge ttd for a post merge hardfork')
-    }
-
-    st.end()
-  })
-
-  t.test(
-    'should throw if encounters a double ttd hardfork specification',
-    function (st: tape.Test) {
-      const c = new Common({ chain: Chain.Sepolia })
-      // Add the ttd to mergeForkIdTransition which occurs post merge in sepolia
-      c.hardforks().filter((hf) => hf.name === 'mergeForkIdTransition')[0]!['ttd'] =
-        '17000000000000000'
-
-      try {
-        c.setHardforkByBlockNumber(1735371)
-        st.fail('should have thrown as two hardforks with ttd specified')
-      } catch (error) {
-        st.pass('throws error as two hardforks with ttd specified')
-      }
-      st.end()
-    }
-  )
-
   t.test(
     'should generate expected hash with london block zero and base fee per gas defined',
     async (st) => {
